--- src/SkillSystem.Aplication/Services/ITokenService.cs
@@ -1,12 +1,0 @@
-﻿using SkillSystem.Aplication.Dtos;
-using System.Security.Claims;
-
-namespace SkillSystem.Aplication.Services;
-
-public interface ITokenService
-{
-    public string GenerateToken(UserGetDto userDto);
-    string GenerateRefreshToken();
-    ClaimsPrincipal? GetPrincipalFromExpiredToken(string token);
-    public string RemoveRefreshTokenAsync(string token);
-}
